Updating for  netcore-3.1
