--- Mongo.Repository.sln
@@ -1,22 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Mongo.Repository", "Mongo.Repository.xproj", "{A5D0012C-6301-4C19-992B-DC4707725180}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A5D0012C-6301-4C19-992B-DC4707725180}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A5D0012C-6301-4C19-992B-DC4707725180}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A5D0012C-6301-4C19-992B-DC4707725180}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A5D0012C-6301-4C19-992B-DC4707725180}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
