+++ MongoDB.Context.Tests/UnitTest1.cs
@@ -1,0 +1,10 @@
+using Xunit;
+
+namespace MongoDB.Context.Tests {
+  public class UnitTest1 {
+    [Fact]
+    public void Test1() {
+
+    }
+  }
+}
