+++ MongoDB.Context.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29911.84
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MongoDB.Context", "MongoDB.Context\MongoDB.Context.csproj", "{5C27F863-53C6-419C-B91F-FCC78214955D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MongoDB.Context.Tests", "MongoDB.Context.Tests\MongoDB.Context.Tests.csproj", "{74F1A4D9-2EDC-4324-91FA-1F44E70C8844}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{5C27F863-53C6-419C-B91F-FCC78214955D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5C27F863-53C6-419C-B91F-FCC78214955D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5C27F863-53C6-419C-B91F-FCC78214955D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5C27F863-53C6-419C-B91F-FCC78214955D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{74F1A4D9-2EDC-4324-91FA-1F44E70C8844}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{74F1A4D9-2EDC-4324-91FA-1F44E70C8844}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{74F1A4D9-2EDC-4324-91FA-1F44E70C8844}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{74F1A4D9-2EDC-4324-91FA-1F44E70C8844}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {77A4E943-9696-40BE-9C45-D87A773154D0}
+	EndGlobalSection
+EndGlobal
