+++ MongoDB.Context/Interfaces/IMongoDbDocument.cs
@@ -1,0 +1,14 @@
+﻿using MongoDB.Bson.Serialization.Attributes;
+using System;
+
+namespace MongoDB.Context.Interfaces {
+
+  public interface IMongoDbDocument {
+    [BsonElement("id"), BsonId]
+    string ID { get; set; }
+
+    [BsonElement("modifiedAt")]
+    DateTime ModifiedAt { get; set; }
+  }
+
+}
