+++ MongoDB.Context/MongoDbContextOptions.cs
@@ -1,0 +1,4 @@
+﻿namespace MongoDB.Context {
+  public class MongoDbContextOptions {
+  }
+}
