--- project.json
@@ -1,17 +1,0 @@
-{
-  "version": "1.0.0-*",
-
-  "dependencies": {
-    "MongoDB.Driver": "2.3.0",
-    "NETStandard.Library": "1.6.0",
-    "Newtonsoft.Json": "9.0.1"
-  },
-
-  "frameworks": {
-    "net45": {
-      "frameworkAssemblies": {
-        "System.Configuration": "4.0.0.0"
-      }
-    }
-  }
-}
